Bring numeric.c and merge all numeric code
