--- include/pgduckdb/types/decimal.hpp
@@ -1,130 +1,0 @@
-#pragma once
-
-#include <cmath>
-#include "duckdb.hpp"
-#include "duckdb/common/extra_type_info.hpp"
-#include "pgduckdb/pg/numeric.hpp"
-
-#include "pgduckdb/utility/cpp_only_file.hpp" // Must be last include.
-
-namespace pgduckdb {
-
-struct NumericAsDouble : public duckdb::ExtraTypeInfo {
-	// Dummy struct to indicate at conversion that the source is a Numeric
-public:
-	NumericAsDouble() : ExtraTypeInfo(duckdb::ExtraTypeInfoType::INVALID_TYPE_INFO) {
-	}
-};
-
-// FIXME: perhaps we want to just make a generic ExtraTypeInfo that holds the Postgres type OID
-struct IsBpChar : public duckdb::ExtraTypeInfo {
-public:
-	IsBpChar() : ExtraTypeInfo(duckdb::ExtraTypeInfoType::INVALID_TYPE_INFO) {
-	}
-};
-
-using duckdb::hugeint_t;
-
-struct DecimalConversionInteger {
-	static int64_t
-	GetPowerOfTen(idx_t index) {
-		static const int64_t POWERS_OF_TEN[] {1,
-		                                      10,
-		                                      100,
-		                                      1000,
-		                                      10000,
-		                                      100000,
-		                                      1000000,
-		                                      10000000,
-		                                      100000000,
-		                                      1000000000,
-		                                      10000000000,
-		                                      100000000000,
-		                                      1000000000000,
-		                                      10000000000000,
-		                                      100000000000000,
-		                                      1000000000000000,
-		                                      10000000000000000,
-		                                      100000000000000000,
-		                                      1000000000000000000};
-		if (index >= 19) {
-			throw duckdb::InternalException("DecimalConversionInteger::GetPowerOfTen - Out of range");
-		}
-		return POWERS_OF_TEN[index];
-	}
-
-	template <class T>
-	static T
-	Finalize(const NumericVar &, T result) {
-		return result;
-	}
-};
-
-struct DecimalConversionHugeint {
-	static hugeint_t
-	GetPowerOfTen(idx_t index) {
-		static const hugeint_t POWERS_OF_TEN[] {
-		    hugeint_t(1),
-		    hugeint_t(10),
-		    hugeint_t(100),
-		    hugeint_t(1000),
-		    hugeint_t(10000),
-		    hugeint_t(100000),
-		    hugeint_t(1000000),
-		    hugeint_t(10000000),
-		    hugeint_t(100000000),
-		    hugeint_t(1000000000),
-		    hugeint_t(10000000000),
-		    hugeint_t(100000000000),
-		    hugeint_t(1000000000000),
-		    hugeint_t(10000000000000),
-		    hugeint_t(100000000000000),
-		    hugeint_t(1000000000000000),
-		    hugeint_t(10000000000000000),
-		    hugeint_t(100000000000000000),
-		    hugeint_t(1000000000000000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(10),
-		    hugeint_t(1000000000000000000) * hugeint_t(100),
-		    hugeint_t(1000000000000000000) * hugeint_t(1000),
-		    hugeint_t(1000000000000000000) * hugeint_t(10000),
-		    hugeint_t(1000000000000000000) * hugeint_t(100000),
-		    hugeint_t(1000000000000000000) * hugeint_t(1000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(10000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(100000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(1000000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(10000000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(100000000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(1000000000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(10000000000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(100000000000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(1000000000000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(10000000000000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(100000000000000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(1000000000000000000),
-		    hugeint_t(1000000000000000000) * hugeint_t(1000000000000000000) * hugeint_t(10),
-		    hugeint_t(1000000000000000000) * hugeint_t(1000000000000000000) * hugeint_t(100)};
-		if (index >= 39) {
-			throw duckdb::InternalException("DecimalConversionHugeint::GetPowerOfTen - Out of range");
-		}
-		return POWERS_OF_TEN[index];
-	}
-
-	static hugeint_t
-	Finalize(const NumericVar &, hugeint_t result) {
-		return result;
-	}
-};
-
-struct DecimalConversionDouble {
-	static double
-	GetPowerOfTen(idx_t index) {
-		return pow(10, double(index));
-	}
-
-	static double
-	Finalize(const NumericVar &numeric, double result) {
-		return result / GetPowerOfTen(numeric.dscale);
-	}
-};
-
-} // namespace pgduckdb
