+++ src/pgduckdb_detoast.cpp
@@ -1,6 +1,7 @@
 #include "duckdb.hpp"
 
 #include "pgduckdb/pgduckdb_types.hpp"
+#include "pgduckdb/pgduckdb_utils.hpp"
 
 extern "C" {
 #include "postgres.h"
