+++ src/scan/heap_reader.cpp
@@ -2,6 +2,7 @@
 
 #include "pgduckdb/scan/heap_reader.hpp"
 #include "pgduckdb/pgduckdb_types.hpp"
+#include "pgduckdb/pgduckdb_utils.hpp"
 
 extern "C" {
 #include "postgres.h"
